checker et faire des vrai fonction

+++ src/Obligation.kt
@@ -1,5 +1,5 @@
 interface Obligation {
-    fun tax_charge():Int
-    fun max_Donation():Int
-    fun help_hours():Int
+    fun tax_charge():Double
+    fun max_Donation():Double
+    fun help_hours(): Int
 }
